// src/api-spec/protobuf/gen/js/tdex/v2/types_pb.d.ts
export interface Fee {
    /**
     * Percentage fee on both assets of the market in basis point.
     *
     * @generated from protobuf field: tdex.v2.MarketFee percentage_fee = 1;
     */
    percentageFee?: MarketFee;
    /**
     * Fixed fee on both assets of the market in satoshi.
     *
     * @generated from protobuf field: tdex.v2.MarketFee fixed_fee = 2;
     */
    fixedFee?: MarketFee;
}
/**
 * @generated from protobuf message tdex.v2.MarketFee
 */
export interface MarketFee {
    /**
     * @generated from protobuf field: int64 base_asset = 1;
     */
    baseAsset: string;
    /**
     * @generated from protobuf field: int64 quote_asset = 2;
     */
    quoteAsset: string;
}
/**
 * @generated from protobuf message tdex.v2.Balance
 */
export interface Balance {
    /**
     * @generated from protobuf field: uint64 base_amount = 1;
     */
    baseAmount: string;
    /**
     * @generated from protobuf field: uint64 quote_amount = 2;
     */
    quoteAmount: string;
}
/**
 * @generated from protobuf message tdex.v2.Market
 */
export interface Market {
    /**
     * @generated from protobuf field: string base_asset = 1;
     */
    baseAsset: string;
    /**
     * @generated from protobuf field: string quote_asset = 2;
     */
    quoteAsset: string;
}
/**
 * @generated from protobuf message tdex.v2.MarketWithFee
 */
export interface MarketWithFee {
    /**
     * @generated from protobuf field: tdex.v2.Market market = 1;
     */
    market?: Market;
    /**
     * @generated from protobuf field: tdex.v2.Fee fee = 2;
     */
    fee?: Fee;
}
/**
 * @generated from protobuf message tdex.v2.Price
 */
export interface Price {
    /**
     * @generated from protobuf field: double base_price = 1;
     */
    basePrice: number;
    /**
     * @generated from protobuf field: double quote_price = 2;
     */
    quotePrice: number;
}
/**
 * @generated from protobuf message tdex.v2.Preview
 */
export interface Preview {
    /**
     * The price of the market.
     *
     * @generated from protobuf field: tdex.v2.Price price = 1;
     */
    price?: Price;
    /**
     * The fees of the market.
     *
     * @generated from protobuf field: tdex.v2.Fee fee = 2;
     */
    fee?: Fee;
    /**
     * The previewd amount (fees excluded).
     *
     * @generated from protobuf field: uint64 amount = 3;
     */
    amount: string;
    /**
     * The asset of the previewed amount (fees excluded).
     *
     * @generated from protobuf field: string asset = 4;
     */
    asset: string;
    /**
     * The previewed fee amount,
     *
     * @generated from protobuf field: uint64 fee_amount = 5;
     */
    feeAmount: string;
    /**
     * The asset of the previewed fee amount,
     *
     * @generated from protobuf field: string fee_asset = 6;
     */
    feeAsset: string;
}
/**
 * @generated from protobuf enum tdex.v2.TradeType
 */
export declare enum TradeType {
    /**
     * @generated from protobuf enum value: TRADE_TYPE_BUY = 0;
     */
    BUY = 0,
    /**
     * @generated from protobuf enum value: TRADE_TYPE_SELL = 1;
     */
    SELL = 1
}
// ...
declare class Fee$Type extends MessageType<Fee> {
    constructor();
    create(value?: PartialMessage<Fee>): Fee;
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Fee): Fee;
    internalBinaryWrite(message: Fee, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter;
}
/**
 * @generated MessageType for protobuf message tdex.v2.Fee
 */
export declare const Fee: Fee$Type;
// ...
export declare const MarketFee: MarketFee$Type;
declare class Balance$Type extends MessageType<Balance> {
    constructor();
    create(value?: PartialMessage<Balance>): Balance;
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Balance): Balance;
    internalBinaryWrite(message: Balance, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter;
}
/**
 * @generated MessageType for protobuf message tdex.v2.Balance
 */
export declare const Balance: Balance$Type;
declare class Market$Type extends MessageType<Market> {
    constructor();
    create(value?: PartialMessage<Market>): Market;
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Market): Market;
    internalBinaryWrite(message: Market, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter;
}
/**
 * @generated MessageType for protobuf message tdex.v2.Market
 */
export declare const Market: Market$Type;
declare class MarketWithFee$Type extends MessageType<MarketWithFee> {
    constructor();
    create(value?: PartialMessage<MarketWithFee>): MarketWithFee;
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: MarketWithFee): MarketWithFee;
    internalBinaryWrite(message: MarketWithFee, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter;
}
/**
 * @generated MessageType for protobuf message tdex.v2.MarketWithFee
 */
export declare const MarketWithFee: MarketWithFee$Type;
declare class Price$Type extends MessageType<Price> {
    constructor();
    create(value?: PartialMessage<Price>): Price;
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Price): Price;
    internalBinaryWrite(message: Price, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter;
}
/**
 * @generated MessageType for protobuf message tdex.v2.Price
 */
export declare const Price: Price$Type;
declare class Preview$Type extends MessageType<Preview> {
    constructor();
    create(value?: PartialMessage<Preview>): Preview;
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Preview): Preview;
    internalBinaryWrite(message: Preview, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter;
}
/**
 * @generated MessageType for protobuf message tdex.v2.Preview
 */
export declare const Preview: Preview$Type;